FMJ fork: Bot AI system, expedition overhaul, mobile CSS, merchant exchange
Bot AI (365 bots live):
- ROI-based decision engine (DOIR) replacing fixed build orders
- Full combat pipeline: espionage, intel, attack, grudge system
- Win/loss adaptation, resource pressure, combat chat
- Debris harvesting, fleet save, noob protection
- Personality system: raider, balanced, turtle, passive

Expedition system rewrite:
- Per-player shuffled deck (1000 cards, guaranteed distribution)
- Scaling percentage ship finds based on fleet size
- Fleet-value resource loot system
- System depletion (anti-spam), Astrophysics limits

Mobile CSS:
- Full responsive design for all game pages
- Hamburger menu, card layouts, touch-friendly inputs

Merchant exchange:
- Resource-to-resource trading with DM cost
- Two-step flow with live preview

Bug fixes:
- Fleet cancel handler, colonization escorts, storage capacity
- Login sessions, messenger SQL injection, shipyard types
- Galaxy AJAX navigation

+++ legacy/app/Http/Controllers/Game/ShipyardController.php
@@ -433,7 +433,7 @@ class ShipyardController extends BaseController
             $buildable['deuterium'] = floor($this->resources_consumed['deuterium'] / $price_deuterium);
         }
 
-        return max(min($buildable), 0);
+        return (int) max(min($buildable), 0);
     }
 
     private function getMaxBuildableItemsBySystemLimit(): int
